Updated logic for updating form in DSL
